e2e-tests: restructure rest tests
