+++ crates/sui-e2e-tests/tests/rest/main.rs
@@ -1,0 +1,5 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+mod execute;
+mod resolve;
